refactor: align orderProgressBar module structure
