--- apps/cowswap-frontend/src/modules/orderProgressBar/OrderProgressBar/styled.ts
@@ -1,13 +1,0 @@
-import styled from 'styled-components/macro'
-
-export const DebugPanel = styled.div`
-  position: fixed;
-  bottom: 10px;
-  right: 10px;
-  background-color: rgba(0, 0, 0, 0.7);
-  color: white;
-  padding: 10px;
-  border-radius: 5px;
-  z-index: 1000;
-`
-
