+++ apps/cowswap-frontend/src/modules/orderProgressBar/helpers.ts
