+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/OrderProgressBar/index.cosmos.tsx
@@ -7,8 +7,8 @@ import styled from 'styled-components/macro'
 
 import { Order } from 'legacy/state/orders/actions'
 
-import { getOrderMock } from '../../../mocks/orderMock'
-import { OrderProgressBarProps } from '../types'
+import { getOrderMock } from '../../../../mocks/orderMock'
+import { OrderProgressBarProps } from '../../types'
 
 import { OrderProgressBar } from './index'
