+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/RenderProgressTopSection.tsx
@@ -4,12 +4,12 @@ import { getRandomInt } from '@cowprotocol/common-utils'
 
 import { AnimatePresence, motion } from 'framer-motion'
 
+import { ProgressTopSection } from './ProgressTopSection'
+import { OrderIntent } from './steps/OrderIntent'
 import * as styledEl from './styled'
 
-import { OrderProgressBarProps } from '../../types'
 import { CHAIN_SPECIFIC_BENEFITS, SURPLUS_IMAGES } from '../constants'
-import { ProgressTopSection } from '../pure/ProgressTopSection'
-import { OrderIntent } from '../steps/OrderIntent'
+import { OrderProgressBarProps } from '../types'
 
 export function RenderProgressTopSection({
   stepName,
