+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/StepComponent.tsx
