+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/StepsWrapper.tsx
