+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/TopSections.tsx
@@ -12,12 +12,12 @@ import SVG from 'react-inlinesvg'
 import { Textfit } from 'react-textfit'
 
 import { ProgressImageWrapper } from './ProgressImageWrapper'
+import AnimatedTokens from './steps/AnimatedToken'
+import { CircularCountdown } from './steps/CircularCountdown'
 import * as styledEl from './styled'
 
-import { OrderProgressBarProps } from '../../types'
 import { truncateWithEllipsis } from '../helpers'
-import AnimatedTokens from '../steps/AnimatedToken'
-import { CircularCountdown } from '../steps/CircularCountdown'
+import { OrderProgressBarProps } from '../types'
 
 interface BaseTopSectionProps {
   stepName: OrderProgressBarProps['stepName']
