+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/steps/AnimatedToken.tsx
