+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/steps/styled.ts
@@ -4,11 +4,9 @@ import { ButtonPrimary, Font, LinkStyledButton, Media, UI } from '@cowprotocol/u
 
 import styled, { css, keyframes } from 'styled-components/macro'
 
-export { Description, ProgressImageWrapper } from '../sharedStyled'
+export { Description, ProgressImageWrapper } from '../../sharedStyled'
 
 import { CancelButton as CancelButtonOriginal } from 'common/pure/CancelButton'
-
-import { ProgressImageWrapper } from '../sharedStyled'
 
 const progressAnimation = (startAt: number, end: number) => {
   const start = end - startAt
