+++ apps/cowswap-frontend/src/modules/orderProgressBar/pure/styled.ts
