+++ apps/cowswap-frontend/src/modules/orderProgressBar/sharedStyled.ts
